+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/authenticator/internal/LDAPAuthResult.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.authenticator.ldap;
+package com.liferay.portal.ldap.authenticator.internal;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.StringPool;
